마우스 휠 조작 버그 수정 및 세분화

+++ YaMang/YaMangClient/YaMangClient/YaMangClient/MainWindow.cpp
@@ -155,7 +155,7 @@ LRESULT MainWindow::HandleMessage( UINT uMsg, WPARAM wParam, LPARAM lParam ) con
 
 		case WM_MOUSEWHEEL:
 		{
-			CameraController::GetInstance()->ChangeMouseZoomStatus(HIWORD(wParam));
+			CameraController::GetInstance()->ChangeMouseZoomStatus( static_cast<short>( HIWORD( wParam ) ) );
 		}
 			return 0;
